Separated the Vue components into files
