+++ package.json
@@ -4,7 +4,7 @@
   "description": "A adiciona funcionalidades e melhora a usabilidade do LegalOne. Esta é uma ferramenta open-source desenvolvida pela comunidade e não possui ligação com a funcionalidade com a Thomson Reuters</b>.",
   "main": "main.entry.js",
   "scripts": {
-    "watch": "webpack --watch",
+    "watch": "webpack --progress --watch",
     "build": "cross-env NODE_ENV=production webpack --progress"
   },
   "repository": {
